+++ src/lib.rs
@@ -32,7 +32,7 @@ mod util;
 mod vec;
 
 pub use arena::ArenaStats;
-pub use arena::{Arena, Checkpoint};
+pub use arena::{Arena, ArenaBox, Checkpoint};
 pub use util::{Transaction, TxnDiff, TxnStatus};
 pub use vec::{ArenaVec, ArenaVecIntoIter, TryReserveError};